+++ src/layouts/MainLayout.js
@@ -11,10 +11,11 @@ const MainLayout = ({ children }) => {
           content="REPLACE_WITH_YOUR_PROJECT_DESCRIPTION"
         />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <meta name="theme-color" content="#F1F0EC" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <NavBar />
-      <div className="flex grow">{children}</div>
+      <div className="flex grow bg-background text-foreground">{children}</div>
     </>
   );
 };